Draw Flow Diagram of a C Function

Draw Flow Diagram of a C Function

build_plantuml_code_from_c(x, fun = NULL, out = NULL)

Arguments

x
url of the file

fun
function as a string, if missing, build_plantuml_code_from_c() will name the possible choices, and return invisibly a named list of the code of the functions.