Embed chart in roxygen doc

Include a call `r_flow::flow_embed(...)` in your doc and a diagram will be included.

flow_embed(call, name, width = 1, alt = name)

Arguments

call

A call to a flow function, prefixed with flow::

width

width, relative if < 1, pixels otherwise

alt

alt text

file

A name for the png file that will be created under 'man/figures', without extension.

Value

Called for side effects, should only be used in roxygen doc

Details

As with images in general the image might not be visible when viewing temp doc with the devtools workflow.

Don't forget to add flow to Suggests in your DESCRIPTION file.

We don't monitor files created under 'man/figures', so if you remove a diagram from the doc make sure to also remove it from the folder.

We also don't overwrite created files, so we don't slow down the documentation process, so if you want to print a different diagram for the same name remove the file first.