View function as flow chart

flow_view() shows the code of a function as a flow diagram, flow_run() runs a call and draws the logical path taken by the code.

flow_view(
  x,
  prefix = NULL,
  truncate = NULL,
  nested_fun = NULL,
  swap = TRUE,
  narrow = FALSE,
  code = TRUE,
  out = NULL,
  svg = FALSE,
  engine = c("nomnoml", "plantuml"),
  engine_opts = getOption("flow.engine_opts")
)

flow_run(
  x,
  prefix = NULL,
  truncate = NULL,
  swap = TRUE,
  code = TRUE,
  out = NULL,
  svg = FALSE,
  browse = FALSE,
  show_passes = FALSE,
  engine_opts = getOption("flow.engine_opts")
)

Arguments

x
A call, a function, or a path to a script

prefix
prefix to use for special comments in our code used as block headers, must start with "#", several prefixes can be provided

truncate
maximum number of characters to be printed per line

nested_fun
if not NULL, the index or name of the function definition found in x that we wish to inspect

swap
whether to change var <- if(cond) expr into if(cond) var <- expr so the diagram displays better

narrow
TRUE makes sure the diagram stays centered on one column (they'll be longer but won't shift to the right)

code
Whether to display the code in code blocks or only the header, to be more compact, if NA, the code will be displayed only if no header is defined by special comments.

out
a path to save the diagram to. Special values "html", "htm", "png", "pdf", "jpg" and "jpeg" can be used to export the objec to a temp file of the relevant format and open it, if a regular path is used the format will be guessed from the extension.

svg
only for default or html output, whether to use svg rendering, rendering is more robust without it, but it makes text selectable on output which is sometimes useful

engine
Either "nomnoml" (default) or "plantuml" (experimental), if the latter, arguments prefix, narrow, and code

engine_opts
A named list.

In the most general case engine_opts is a list containing optional sub-lists named "nomnoml", "plantuml", and "htmlwidgets".

The items of the nomnoml sub-list will be forwarded to nomnoml:::build_nomnoml_code(), unexported but documented in ?build_nomnoml_code, in order to alter nomnoml diagrams. For instance fontSize, direction = "right" to have a left to right diagram, ranker = "longest-path" to see have exit points at the bottom...

The items of the plantuml sub-list will be forwarded to plantuml:::plot.plantuml(), , in order to alter plantuml diagrams. see ?plantuml:::plot.plantuml and ?plantuml::plantuml_run

The items of the htmlwidgets sub-list will be forwarded to htmlwidgets::createWidget() or htmlwidgets::saveWidget(), which are used by the nomnoml engine.

If this structure is not found, the items of engine_opts will be forwarded to nomnoml:::build_nomnoml_code() if the engine is "nomnoml", and plantuml:::plot.plantuml if the engine is "plantuml"

browse
whether to debug step by step (block by block), can also be a vector of block ids, in this case browser() calls will be inserted at the start of these blocks

show_passes
label the edges with the number of passes